+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## Current
 
 * fix: Shutdown websocket when the tcp client terminates. #38
-* fix: Removed `--reconnects` option as the websocket connection should not be reattempted after it is established as it causes problems with different clients. #40
+* fix: Mark `--reconnects` option as deprecated, it will be removed in the next major release. The reconnect logic websocket has been removed as attempting a reconnect after the client has been connected once causes problems with different servers. #40
 
 ## 2.0.4